requirements.txt
tensorflow[and-cuda]
nvidia-tensorrt
tensorflow-datasets
matplotlib
ipykernel
numpy
pandas
scipy
opencv-python
python-telegram-bot
